fix: Изменил типы сумм на BigDecimal устранив отображение сумм в отчетах
fix: добавлена проверка на ввод сумм, теперь корректно добавляются договора с целыми числами
refactor: Описал dto, добавил валидацию полей
delete: Удалены неиспользуемые dto and view

+++ src/main/java/ru/hometask/dto/UpdateContractDto.java
@@ -16,7 +16,7 @@ public class UpdateContractDto {
 
     private Long clientId;
 
-    private double amount;
+    private BigDecimal amount;
 
     private LocalDateTime dateOfIssue;
